refactor(canvas-workspace): consolidate harness surfaces under harness/
Physical unification per the five-surface model: the workspace's repo-harness
surfaces had grown scattered (docs/, skills/, a bare harness/ driver). They now
share one container so an agent finds all of them in one place, aligned with
packages/engine/harness/:

  harness/
    knowledge/            <- docs/conventions/ + main-domain-modules.md
                             + renderer-surfaces.md + plugin-node-mf2.md
    skills/               <- skills/ (canvas-harness, canvas-onboard-harness)
    tools/driver/         <- the headless-Electron driver (was harness/{cli.mjs,src,README.md})
    validate/             <- unchanged (already here)

docs/ keeps only project records (perf analyses, roadmaps), which are not
harness surfaces. All 30 files moved via git rename (history preserved).

Wiring updated so nothing breaks:
- driver self-location: config.mjs APP_DIR depth +2 (src -> driver -> tools ->
  harness -> app root); verified APP_DIR still resolves to the app root and the
  moved vitest suite passes 5/5 at the new path.
- package.json `harness` script -> harness/tools/driver/cli.mjs.
- LOAD-BEARING: scripts/perf/{report,run-scenarios}.mjs import driver internals
  (session/cdp/utils.mjs) and spawn cli.mjs — repointed to harness/tools/driver/;
  this is the perf CI path (.github/workflows/perf.yml), so a stale ref would
  have gone red on the next PR.
- AGENTS.md: nav table repointed, harness-layout section replaces the old
  two-meanings note (the directory is no longer product-only), Key Files entry
  rewritten; README.md downstream links repointed.
- validate/validation.yaml product-harness rule now globs harness/tools/driver/**.
- root harness/DESIGN.md: the canvas example rewritten from "don't overload the
  driver dir" into the worked precedent for migrating a single-mechanism harness/
  into the full five-surface layout.

check-harness: harnessGaps 0. Full straggler sweep clean.
