+++ apps/canvas-workspace/harness/knowledge/plugin-node-mf2.md
