+++ apps/canvas-workspace/harness/knowledge/renderer-surfaces.md
