+++ apps/canvas-workspace/harness/skills/canvas-harness/agents/openai.yaml
