+++ apps/canvas-workspace/harness/skills/canvas-onboard-harness/SKILL.md
