+++ apps/canvas-workspace/harness/skills/canvas-onboard-harness/agents/openai.yaml
