+++ apps/canvas-workspace/harness/tools/driver/src/config.mjs
@@ -1,7 +1,9 @@
 import { dirname, join, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
-export const APP_DIR = resolve(dirname(fileURLToPath(import.meta.url)), '..', '..');
+// src/ -> driver/ -> tools/ -> harness/ -> app root (this file self-locates;
+// keep the depth in sync if the driver moves again)
+export const APP_DIR = resolve(dirname(fileURLToPath(import.meta.url)), '..', '..', '..', '..');
 export const HARNESS_DIR = join(APP_DIR, '.harness');
 export const CURRENT_SESSION_PATH = join(HARNESS_DIR, 'current-session.json');
 export const DIST_MAIN = join(APP_DIR, 'dist', 'main', 'index.js');
