+++ apps/canvas-workspace/harness/tools/driver/src/renderer.mjs
