+++ apps/canvas-workspace/harness/tools/driver/src/screenshot.mjs
