+++ apps/canvas-workspace/perf/README.md
@@ -54,9 +54,9 @@ run standalone without the flag) just omits the section — no error.
 ```bash
 pnpm --filter canvas-workspace setup:electron    # only if the binary download was skipped/blocked
 pnpm --filter canvas-workspace build
-node harness/cli.mjs start --profile temp --headless   # display-less Linux; omit --headless on a desktop (see harness/README.md)
+node harness/tools/driver/cli.mjs start --profile temp --headless   # display-less Linux; omit --headless on a desktop (see harness/tools/driver/README.md)
 pnpm --filter canvas-workspace perf:scenarios -- --seed-nodes 100
-node harness/cli.mjs close --cleanup
+node harness/tools/driver/cli.mjs close --cleanup
 ```
 
 ## Dashboard (from real data)
