+++ harness/DESIGN.md
@@ -100,7 +100,7 @@ Use these defaults:
 - Add `harness/spec/` only when normative design is needed for a feature, protocol, migration, or architectural boundary.
 - Add `harness/skills/` only for stable local action protocols that are not runtime task skills.
 
-If a workspace already has `harness/` for a specific mechanism, treat that existing directory as the relevant surface until there is enough pressure to split it. For example, `apps/canvas-workspace/harness/` is currently a real Electron operation tool; do not mix unrelated Knowledge or Skills files into it casually. If it later needs a full workspace harness container, migrate deliberately instead of overloading the existing tool layout.
+If a workspace already has `harness/` for a specific mechanism, treat that existing directory as the relevant surface until there is enough pressure to split it, then migrate deliberately into the full five-surface layout rather than overloading the existing sub-layout. `apps/canvas-workspace/harness/` did exactly this (2026-07-07): it began as a bare Electron operation tool directly under `harness/`, and once the workspace's Knowledge (conventions and domain maps, previously under the app's docs directory) and Skills (the SKILL.md procedures) were ready to co-locate, the driver was moved down into `harness/tools/driver/` and the other surfaces took their named slots (`harness/knowledge/`, `harness/skills/`, alongside the pre-existing `harness/validate/`). The driver keeps working because it self-locates via `import.meta.url` and is invoked through the `harness` package script, not a hardcoded path.
 
 Runtime artifact directories such as `.harness/` are not harness knowledge. They hold generated state, screenshots, logs, or temporary homes and should not become documentation or validation sources of truth.
